requirements.txt
ms-swift==3.10.2
transformers==4.57.1
qwen_vl_utils>=0.0.14
trl[vllm]
math_verify
wandb
weave
packaging
ninja
deepspeed
vllm==0.11.0
msgspec
torchvision
decord
pandas
tqdm
Pillow
flash-attn